Updating RAM usage calculations

+++ libraries/chain/include/eosio/chain/backing_store/db_key_value_sec_lookup.hpp
@@ -21,7 +21,7 @@ namespace eosio { namespace chain { namespace backing_store {
          EOS_ASSERT( pp.value_size == 0, db_rocksdb_invalid_operation_exception,
                      "Payload should not have anything more than the payer");
       }
-      int64_t overhead() { return config::billable_size_v<index64_object>;}
+      uint64_t overhead() { return config::billable_size_v<index64_object>;}
    };
 
    template<>
@@ -36,7 +36,7 @@ namespace eosio { namespace chain { namespace backing_store {
          EOS_ASSERT( pp.value_size == 0, db_rocksdb_invalid_operation_exception,
                      "Payload should not have anything more than the payer");
       }
-      int64_t overhead() { return config::billable_size_v<index128_object>;}
+      uint64_t overhead() { return config::billable_size_v<index128_object>;}
    };
 
    template<>
@@ -51,7 +51,7 @@ namespace eosio { namespace chain { namespace backing_store {
          EOS_ASSERT( pp.value_size == 0, db_rocksdb_invalid_operation_exception,
                      "Payload should not have anything more than the payer");
       }
-      int64_t overhead() { return config::billable_size_v<index256_object>;}
+      uint64_t overhead() { return config::billable_size_v<index256_object>;}
    };
 
    template<>
@@ -73,7 +73,7 @@ namespace eosio { namespace chain { namespace backing_store {
          // see note on value(...) method above
          memcpy(&sec_key, pp.value, pp.value_size);
       }
-      int64_t overhead() { return config::billable_size_v<index_double_object>;}
+      uint64_t overhead() { return config::billable_size_v<index_double_object>;}
    };
 
    template<>
@@ -96,7 +96,7 @@ namespace eosio { namespace chain { namespace backing_store {
          // see note on value(...) method above
          memcpy(&sec_key, pp.value, pp.value_size);
       }
-      int64_t overhead() { return config::billable_size_v<index_long_double_object>;}
+      uint64_t overhead() { return config::billable_size_v<index_long_double_object>;}
    };
 
    template<typename SecondaryKey, typename Session>
@@ -193,8 +193,8 @@ namespace eosio { namespace chain { namespace backing_store {
          }
 
          if( key_store.payer != payer ) {
-            context.update_db_usage( key_store.payer, -(helper.overhead()), backing_store::db_context::secondary_update_rem_trace(context.get_action_id(), std::string(event_id)) );
-            context.update_db_usage( payer, +(helper.overhead()), backing_store::db_context::secondary_update_add_trace(context.get_action_id(), std::move(event_id)) );
+            context.update_db_usage( key_store.payer, -(static_cast<int64_t>(helper.overhead())), backing_store::db_context::secondary_update_rem_trace(context.get_action_id(), std::string(event_id)) );
+            context.update_db_usage( payer, +(static_cast<int64_t>(helper.overhead())), backing_store::db_context::secondary_update_add_trace(context.get_action_id(), std::move(event_id)) );
          }
 
          // if the secondary value is different, remove the old key and add the new key
